Implemented Cart Management

+++ MenuItemAPI/views.py
@@ -18,7 +18,7 @@ class MenuItemList(APIView):
     permission_classes = [IsAuthenticated]
     
     def get(self, request, format=None):
-        items = MenuItem.objects.select_related('category').all()
+        items = MenuItem.objects.select_related('category')
         category_name = request.query_params.get('category')
         l_price = request.query_params.get('l_price')
         r_price = request.query_params.get('r_price')
@@ -36,10 +36,13 @@ class MenuItemList(APIView):
         return Response(serialized_items.data, status=status.HTTP_200_OK)
     
     def post(self, request, format=None):
-        serialized_item = MenuItemSerializer(data=request.data)
-        serialized_item.is_valid(raise_exception=True)
-        serialized_item.save()
-        return Response(serialized_item.data, status=status.HTTP_201_CREATED)
+        if request.user.groups.filter(name="Manager").exists():
+            serialized_item = MenuItemSerializer(data=request.data)
+            serialized_item.is_valid(raise_exception=True)
+            serialized_item.save()
+            return Response(serialized_item.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response({'message':'Access Denied'}, status=status.HTTP_403_FORBIDDEN)
 
 class MenuItemDetail(APIView):
     def get(self, request, pk, format=None):
@@ -48,16 +51,23 @@ class MenuItemDetail(APIView):
         return Response(serialized_item.data, status=status.HTTP_200_OK)
 
     def put(self, request, pk, format=None):
-        item = get_object_or_404(MenuItem, pk=pk)
-        serialized_item = MenuItemSerializer(item, data=request.data)
-        serialized_item.is_valid(raise_exception=True)
-        serialized_item.save()
-        return Response(serialized_item.data, status=status.HTTP_201_CREATED)
+        if request.user.groups.filter(name="Manager").exists():
+            item = get_object_or_404(MenuItem, pk=pk)
+            serialized_item = MenuItemSerializer(item, data=request.data)
+            serialized_item.is_valid(raise_exception=True)
+            serialized_item.save()
+            return Response(serialized_item.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response(status=status.HTTP_403_FORBIDDEN)
     
     def delete(self, request, pk, format=None):
-        item = get_object_or_404(MenuItem, pk=pk)
-        item.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.user.groups.filter(name="Manager").exists():
+            item = get_object_or_404(MenuItem, pk=pk)
+            item.delete()
+            return Response(status=status.HTTP_200_OK)
+        else:
+            return Response(status=status.HTTP_403_FORBIDDEN)
+
